[hincrbyfloat] double -> long double

+++ ssdb-1.9.2/src/ssdb/ssdb.h
@@ -80,7 +80,7 @@ public:
 
 	// -1: error, 1: ok, 0: value is not an integer or out of range
 	virtual int hincr(const Bytes &name, const Bytes &key, int64_t by, int64_t *new_val) = 0;
-	virtual int hincrbyfloat(const Bytes &name, const Bytes &key, double by, double *new_val) = 0;
+	virtual int hincrbyfloat(const Bytes &name, const Bytes &key, long double by, long double *new_val) = 0;
 
 	virtual int64_t hsize(const Bytes &name) = 0;
 	virtual int hget(const Bytes &name, const Bytes &key, std::string *val) = 0;
